backend/app/requirements.txt
fastapi
uvicorn
requests
pymysql
sqlalchemy
certifi
httpx
passlib[bcrypt]
pyjwt
python-dotenv
motor
